tweaks

+++ key.go
@@ -2,11 +2,11 @@ package main
 
 var keyChar = []byte("0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
 
-func genKey(n int64) string {
+func genKey(n int) string {
 	if n == 0 {
 		return string(keyChar[0])
 	}
-	l := int64(len(keyChar))
+	l := len(keyChar)
 	s := make([]byte, 20) // FIXME: will overflow. eventually.
 	i := len(s)
 	for n > 0 && i >= 0 {
